pyproject.toml
[project]
name = "finrizz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.0.11",
    "ddgs>=9.6.0",
    "firecrawl-py>=4.3.6",
    "google-genai>=1.39.1",
    "google-generativeai>=0.8.5",
    "lxml-html-clean>=0.4.2",
    "openai>=1.109.1",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.5",
]
